Update attack to utilize state pattern.
