+++ Sprintfinity3902/Interfaces/IPlayerState.cs
@@ -8,5 +8,7 @@ namespace Sprintfinity3902.Interfaces
     {
         ISprite Sprite { get; set; }
         void Move();
+
+        void Attack();
     }
 }
