+++ Sprintfinity3902/States/FacingRightState.cs
@@ -24,5 +24,10 @@ namespace Sprintfinity3902.States
             PlayerCharacter.setCurrentPositionX(Sprite.CurrentPositionX);
         }
 
+        public void Attack()
+        {
+            //Needs to be implemented
+        }
+
     }
 }
